% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugAdverseEvents.R
\name{plotDrugAdverseEvents}
\alias{plotDrugAdverseEvents}
\title{Allow users to visualize drug information against selected variable.}
\usage{
plotDrugAdverseEvents(drugs_and_events, drug_of_interest = "LETAIRIS",
  variable_of_interest)
}
\arguments{
\item{drugs_and_events}{the dataframe that is returned from calling queryOpenFDA in queryOpenFDA.R}

\item{drug_of_interest}{a drug of interest for the user, must be one of the drugs
within the drugs_and_events dataframe}

\item{variable_of_interest}{the variable of interest to plot from the columns
of drugs_and_events dataframe. Should be one of: c("age", "sex", "reaction", "date")}
}
\value{
a ggplot with the variable of interest as the x axis and the count of that variable for
a specific drug on the y axis.
}
\description{
Allow users to request to plot a drug against a patient demographic variable
}
\examples{
drugs_and_events <- queryAdverseEvents(limit = 101)
plotDrugAdverseEvents(drugs_and_events, "LETAIRIS", "age")
}
